More compil compatibility

+++ trunk/src/PotatoEngine/Debug/Renderer.hpp
@@ -43,7 +43,7 @@ namespace Debug
 			void DrawRect(const Vector2f& p1, const Vector2f& p2, const Vector2f& p3, const Vector2f& p4, const sf::Color& color);
 			void DrawRect(const Vector2f* vertices, const sf::Color& color);
 			void DrawPolygon(const Vector2f* vertices, int32 vertexCount, const sf::Color& color);
-			void DrawSolidPolygon(const Vector2f* vertices, int32 vertexCount, const sf::Color& color);
+			void DrawSolidPolygon(const Vector2f* vertices, size_t vertexCount, const sf::Color& color);
 			void DrawCircle(const Vector2f& center, float32 radius, const sf::Color& color);
 			void DrawSolidCircle(const Vector2f& center, float32 radius, const Vector2f& axis, const sf::Color& color);
 			
@@ -59,7 +59,7 @@ namespace Debug
 			void DrawRect(const sf::Vector2f& p1, const sf::Vector2f& p2, const sf::Vector2f& p3, const sf::Vector2f& p4, const sf::Color& color);
 			void DrawRect(const sf::Vector2f* vertices, const sf::Color& color);
 			void DrawPolygon(const sf::Vector2f* vertices, int32 vertexCount, const sf::Color& color);
-			void DrawSolidPolygon(const sf::Vector2f* vertices, int32 vertexCount, const sf::Color& color);
+			void DrawSolidPolygon(const sf::Vector2f* vertices, size_t vertexCount, const sf::Color& color);
 			void DrawCircle(const sf::Vector2f& center, float32 radius, const sf::Color& color);
 			void DrawSolidCircle(const sf::Vector2f& center, float32 radius, const sf::Vector2f& axis, const sf::Color& color);
 			void DrawAABB(const sf::FloatRect& aabb, const sf::Color& color);
